+++ CHANGELOG.md
@@ -9,6 +9,7 @@ New features:
 - Updated the terminal AI tool `get_status` operation to return the operating system and shell type.  Updated the tool
   description to avoid leading to any assumption of running on a Unix-like system.
 - Added a system prompt to provide initial guidance to all LLMs when starting conversations.
+- Add a `get_system_info` operation to the system AI tool.  This provides system and mindspace details.
 
 Bug fixes: